Learning game-like for rounds

+++ evolve_bell.py
@@ -16,11 +16,14 @@ import random
 
 runs_per_net = 25
 simulation_seconds = 60.0
-ngenerations = 10000
+ngenerations = 3000
 
-for i in range(0,10000):
-    if os.path.isfile('./current_network/%d' % i):
-        os.remove('./current_network/%d' % i)
+use_existing_population = True #Load an existing network that is presumably better than nothing
+
+if not use_existing_population:
+    for i in range(0,10000):
+        if os.path.isfile('./current_network/%d' % i):
+            os.remove('./current_network/%d' % i)
 
 # Use the NN network phenotype and the discrete actuator force function.
 def eval_genome(genome, config):
@@ -32,18 +35,23 @@ def eval_genome(genome, config):
         sim = run_bell()  # all the physics in here
 
         if True:   #Conditions for general ringing. Start stood at each stroke I think, but might change that.
-            if runs < int(runs_per_net/2):
-                sim.bell.bell_angle = uniform(np.pi+0.01*sim.bell.stay_angle, np.pi+sim.bell.stay_angle)
+            if True:#runs < int(runs_per_net/2):
+                sim.bell.bell_angle = uniform(np.pi+0.5*sim.bell.stay_angle, np.pi+sim.bell.stay_angle)
                 sim.bell.clapper_angle = sim.bell.bell_angle + sim.bell.clapper_limit - 0.01
             else:
-                sim.bell.bell_angle = uniform(-np.pi-0.01*sim.bell.stay_angle, -np.pi-sim.bell.stay_angle)
+                sim.bell.bell_angle = uniform(-np.pi-0.5*sim.bell.stay_angle, -np.pi -sim.bell.stay_angle)
                 sim.bell.clapper_angle = sim.bell.bell_angle - sim.bell.clapper_limit + 0.01
 
             sim.bell.target_period = uniform(4.6,4.8)
 
-            sim.bell.m_1 = 400
+            sim.bell.m_1 = uniform(390,410)
             sim.bell.m_2 = 0.05*sim.bell.m_1
             sim.bell.stay_break_limit = 0.4
+
+        if np.abs(sim.bell.bell_angle) < 0.5:
+            sim.bell.max_length = 0.0  # max backstroke length
+        else:
+            sim.bell.max_length = sim.bell.radius*(1.0 + 3*np.pi/2 - sim.bell.garter_hole)
 
         # Run the given simulation for up to num_steps time steps.
         fitness = 0.0
@@ -58,12 +66,12 @@ def eval_genome(genome, config):
             sim.bell.pull = force
             sim.step(force)
 
-            strike_limit = 1.0  #1 second out in each direction
+            strike_limit = 0.5  #5 seconds out in each direction to begin with
             #Exit if out of bounds
-            if len(sim.bell.handstroke_accuracy) > 0:
+            if len(sim.bell.handstroke_accuracy) > 1:
                 if np.abs(sim.bell.handstroke_accuracy[-1]) > strike_limit:
                     break
-                if len(sim.bell.backstroke_accuracy) > 0:
+                if len(sim.bell.backstroke_accuracy) > 1:
                     if np.abs(sim.bell.backstroke_accuracy[-1]) > strike_limit:
                         break
                 if sim.bell.stay_hit > 0:
@@ -90,7 +98,17 @@ def run():
     # Load in config file. Will tweak in due course.
     config = neat.Config(neat.DefaultGenome, neat.DefaultReproduction, neat.DefaultSpeciesSet, neat.DefaultStagnation, config_path)
 
-    pop = neat.Population(config)
+    if use_existing_population:
+        with open("population_data/population", "rb") as f:
+            pop_old = pickle.load(f)
+        with open("population_data/species", "rb") as f:
+            species_old = pickle.load(f)
+        with open("population_data/generation", "rb") as f:
+            generation_old = pickle.load(f)
+        initial_state = (pop_old, species_old, generation_old)
+        pop = neat.Population(config, initial_state = initial_state)
+    else:
+        pop = neat.Population(config)
     # These just print some things out. But keep on...
     stats = neat.StatisticsReporter()
     pop.add_reporter(stats)
